refactor: 파이프라인 서비스 예외처리

+++ src/main/java/be15fintomatokatchupbe/campaign/command/application/service/ContractCommandService.java
@@ -138,6 +138,11 @@ public class ContractCommandService {
 
         Pipeline foundPipeline = campaignHelperService.findValidPipeline(request.getPipelineId());
 
+        if(!Objects.equals(foundPipeline.getPipelineStep().getPipelineStepId(), PipelineStepConstants.CONTRACT)){
+            throw new BusinessException(CampaignErrorCode.INVALID_ACCESS);
+        }
+
+
         /* 연관 테이블 지워주기 */
         campaignHelperService.deleteRelationTable(foundPipeline);
 
@@ -186,6 +191,10 @@ public class ContractCommandService {
 
         if(foundPipeline.getPipelineStatus().getPipelineStatusId().equals(PipelineStatusConstants.APPROVED)){
             throw new BusinessException(CampaignErrorCode.APPROVED_PIPELINE_CANNOT_BE_DELETED);
+        }
+
+        if(!Objects.equals(foundPipeline.getPipelineStep().getPipelineStepId(), PipelineStepConstants.CONTRACT)){
+            throw new BusinessException(CampaignErrorCode.INVALID_ACCESS);
         }
 
         // 2. 파이프라인 소프트 딜리트 하기
